requirements.txt
Flask==3.0.0
gunicorn
python-dotenv
requests
requests-oauthlib
